Added integ tests for nonoverlapping case
